+++ src/database/database.ts
@@ -96,4 +96,22 @@ export class SqlSealDatabase {
         const result = await this.db.select(statement, frontmatter)
         return result
     }
+
+    async explain(statement: string, frontmatter: Record<string, unknown>) {
+        const explainResults = await this.db.explainQuery(statement, frontmatter)
+        let strResult = '';
+        const map = new Map<number, number>()
+        const INDENT_INCREASE = 4
+        map.set(0, -INDENT_INCREASE)
+        for (const result of explainResults) {
+            const parent = parseInt((result.parent as unknown as  string) ?? 0, 10)
+            const indent = map.get(parent)! + INDENT_INCREASE
+            for (let i=0;i<indent;i++) {
+                strResult += ' '
+            }
+            strResult += result.detail + "\n"
+            map.set(result.id as number, indent)
+        }
+        return strResult
+    }
 }